GL class draw method now takes array of objects
Decided this was a bit more straightforward than having GL instances
store their own list of objects. Instead, just accept a list and draw
each object in it. This makes the "owner" of the list of objects
clearer. The owner is the caller, since the caller can manipulate the
objects, rotating and translating them, whereas the GL instance is
merely responsible for drawing them.