Automatically calculate placement of the walls
The "left", "right", "top" and "bottom" walls of the game area are now
placed so that, where they're nearest to the viewer, they appear to
touch the edges of the canvas. The "front" wall is placed at the minimum
visible z value, and the "back" wall is placed at the maximum visible z
value.

The idea is that the volume enclosed by the walls should fill the entire
canvas, but not have any regions not visible in the canvas.