Make it easier to set dimensions of the game area
The setPerspective method of GL objects now accepts a desired maxX and
maxY value and scales the perspective matrix so that those values are
the max visible at z near. In effect, these will be the maximum x and y
of the game area. The minimum x and y will be -maxX and -maxY.