Boxes are now made up of Faces
Since I have a "face" class, it makes sense that a box would have a set
of faces, and that the box's polys would get calculated from the faces.

This will now make it easier to look for collisions with boxes. (Since
my "Face" object already had an "intersection" method).