Now a set of bricks that disappears when hit
Gave the boxes representing the bricks have their own
collision-prediction method (and not just the faces that make up the
boxes). This means that the whole box can "react" (by calling its on
collision method and thus disappearing).

Also, if the ball hits two bricks from the same side, it should still
bounce in the opposite direction. Before, I was switching it to the
opposite direction due to one brick, then bouncing it back due to the
other.