Paddle motion now impacts ball motion
Specifically, if the paddle is moving when the ball hits it, the ball
will be deflected in the direction of the paddle's motion.

Now we can randomize the ball's starting position, so doing that too.
(Before, certain starting positions were "unwinnable" - because the ball
would never actually bounce to certain bricks - so we couldn't randomize
the starting position. But now the player can adjust the direction of
the ball if needed.)

Also, did cleanup/clarification of comments.