Fix to setPerspectiveMatrix
Before, setPerspectiveMatrix would make the display look distorted if
the aspect ratio of xMax/yMax didn't match the aspect ratio of the
canvas. We were stretching to fill the entire canvas, which meant that
sometimes we would stretch along the x axis more than along the y, or
vice versa, causing the distortion.

This problem is now fixed - instead of stretching along the x axis to
fill the entire canvas and stretching separately along the y axis to
fill the entire canvas, we now stretch along the x and y axes by the
same amount. If this means that we don't use the entire width or height
of the canvas, that is fine.