Allow levels to adjust xMin and yMin
I was allowing xMax and yMax to be adjusted, but was automatically
setting xMin and yMin. For completeness/generality, let all of these be
set. To make this work, also tweak the scaling/translation of the
perspective matrix so that the game area fits nicely in the viewport,
based on what the xMin/xMax and yMin/yMax values were.