Attempt to predict collisions involving "movers"
For "Movers", adjust the velocity of the incoming object by subtracting
the velocity of the mover before predicting their collision time. This
gets us an approximately correct prediction of the collision time.

However, currently the prediction doesn't work well. One major issue is
that the collision prediction and the initial bounce away from the
collided object both occur within the update method of the ball. If
mover.update is called after ball.update, our predicted collision time
will be incorrect, because it was based on the mover's position at an
earlier time (the time of our last update).  If mover.update is called
before ball.update, the ball will be overlapping the mover when it
bounces away (because the mover will have advanced to a later time), and
we may get stuck inside the mover.

What we really need to do is update ALL objects to the collision time
(and not significantly before or after). THEN update all velocities
based on the collision. THEN figure out when the next collision will
occur. THEN updates to positions can proceed.

This will require some refactoring.