Clarify the "initialize" function in game.js
Initialize serves a bunch of purposes involving setting up a level.

But it had a lot of random things going on for one function and was hard
to understand and summarize. So add a couple of functions - 1)
prepDisplay and 2) refineLevel.

Those functions  1) set up the GLManager based on the level dimensions
(i.e., set canvas dimensions and GL perspective) and 2) create "ball",
"paddle" and "walls" objects (i.e., things that are needed in every
level, that could be set by the level but might not be).

These functions are now called by initialize, rather than all being
lumped together.

Besides that, the only other thing initialize has to do is to set the
canvas to track mouse motion so that the player can move the paddle.

Did some other renaming/reorganizing as well.